--- cpp/include/raft/comms/nccl_helper.hpp
@@ -1,34 +1,0 @@
-/*
- * Copyright (c) 2020, NVIDIA CORPORATION.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-#pragma once
-
-#include <nccl.h>
-
-namespace raft {
-namespace comms {
-inline void ncclUniqueIdFromChar(ncclUniqueId *id, char *uniqueId, int size) {
-  memcpy(id->internal, uniqueId, size);
-}
-
-inline void get_unique_id(char *uid, int size) {
-  ncclUniqueId id;
-  ncclGetUniqueId(&id);
-
-  memcpy(uid, id.internal, size);
-}
-}  // namespace comms
-}  // namespace raft
