+++ python/raft/dask/__init__.py
@@ -13,4 +13,4 @@
 # limitations under the License.
 #
 
-from raft.dask.common.comms import Comms
+from .common.comms import Comms
