+++ python/raft/dask/common/__init__.py
@@ -13,14 +13,14 @@
 # limitations under the License.
 #
 
-from raft.dask.common.comms import Comms
-from raft.dask.common.comms import local_handle
+from .comms import Comms
+from .comms import local_handle
 
-from raft.dask.common.comms_utils import inject_comms_on_handle
-from raft.dask.common.comms_utils import inject_comms_on_handle_coll_only
-from raft.dask.common.comms_utils import perform_test_comms_allreduce
-from raft.dask.common.comms_utils import perform_test_comms_send_recv
-from raft.dask.common.comms_utils import perform_test_comms_allgather
-from raft.dask.common.comms_utils import perform_test_comms_bcast
-from raft.dask.common.comms_utils import perform_test_comms_reduce
-from raft.dask.common.comms_utils import perform_test_comms_reducescatter
+from .comms_utils import inject_comms_on_handle
+from .comms_utils import inject_comms_on_handle_coll_only
+from .comms_utils import perform_test_comms_allreduce
+from .comms_utils import perform_test_comms_send_recv
+from .comms_utils import perform_test_comms_allgather
+from .comms_utils import perform_test_comms_bcast
+from .comms_utils import perform_test_comms_reduce
+from .comms_utils import perform_test_comms_reducescatter
