+++ python/raft/dask/common/comms_utils.pyx
@@ -41,7 +41,7 @@ cdef extern from "raft/comms/std_comms.hpp" namespace "raft::comms":
     cdef cppclass std_comms:
         pass
 
-cdef extern from "raft/comms/comms_helper.hpp" namespace "raft::comms":
+cdef extern from "raft/comms/helper.hpp" namespace "raft::comms":
 
     void build_comms_nccl_ucx(handle_t *handle,
                               ncclComm_t comm,
@@ -54,6 +54,8 @@ cdef extern from "raft/comms/comms_helper.hpp" namespace "raft::comms":
                                ncclComm_t comm,
                                int size,
                                int rank) except +
+
+cdef extern from "raft/comms/test.hpp" namespace "raft::comms":
 
     bool test_collective_allreduce(const handle_t &h, int root) except +
     bool test_collective_broadcast(const handle_t &h, int root) except +
